+++ src/app/utils/boardsFunctions.ts
@@ -28,17 +28,19 @@ export const useBoards = () => {
     category: string
   ) => {
     try {
-      const token = localStorage.getItem("accessToken");
-      const response = await api.get(
-        `/boards?page=${page}&size=${size}&category=${category}`,
-        {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-      setBoards(response.data.boards);
-      setTotalCount(response.data.totalCount);
+      if (typeof window !== "undefined") {
+        const token = localStorage.getItem("accessToken");
+        const response = await api.get(
+          `/boards?page=${page}&size=${size}&category=${category}`,
+          {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+          }
+        );
+        setBoards(response.data.boards);
+        setTotalCount(response.data.totalCount);
+      }
     } catch (error: unknown) {
       if (error instanceof AxiosError) {
         console.error("게시판 불러오기 오류:", error);
@@ -59,15 +61,17 @@ export const useBoards = () => {
   // 게시물 삭제
   const handleDelete = async (boardId: string) => {
     try {
-      const token = localStorage.getItem("accessToken");
-      await api.delete(`/boards/${boardId}`, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-      setBoards(boards.filter((board) => board.id !== boardId));
-      alert("게시물 삭제 완료.");
-      fetchBoards(1, pageSize, category);
+      if (typeof window !== "undefined") {
+        const token = localStorage.getItem("accessToken");
+        await api.delete(`/boards/${boardId}`, {
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
+        });
+        setBoards(boards.filter((board) => board.id !== boardId));
+        alert("게시물 삭제 완료.");
+        fetchBoards(1, pageSize, category);
+      }
     } catch (error) {
       console.error("게시물 삭제 오류:", error);
     }
@@ -76,19 +80,21 @@ export const useBoards = () => {
   // 게시물 생성
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-    const token = localStorage.getItem("accessToken");
     try {
-      await api.post(
-        "/boards",
-        { title, content, category: createCategory },
-        {
-          headers: { Authorization: `Bearer ${token}` },
-        }
-      );
-      alert("게시물 작성 완료.");
-      setEditMode(false);
-      setIsWriting(false);
-      fetchBoards(page, pageSize, category);
+      if (typeof window !== "undefined") {
+        const token = localStorage.getItem("accessToken");
+        await api.post(
+          "/boards",
+          { title, content, category: createCategory },
+          {
+            headers: { Authorization: `Bearer ${token}` },
+          }
+        );
+        alert("게시물 작성 완료.");
+        setEditMode(false);
+        setIsWriting(false);
+        fetchBoards(page, pageSize, category);
+      }
     } catch (error: unknown) {
       if (error instanceof AxiosError) {
         console.error("게시판 작성 오류:", error);
@@ -109,22 +115,24 @@ export const useBoards = () => {
   // 수정된 게시물 제출
   const handleUpdate = async (e: React.FormEvent) => {
     e.preventDefault();
-    const token = localStorage.getItem("accessToken");
     try {
-      await api.put(
-        `/boards/${editBoardId}`,
-        { title, content },
-        {
-          headers: { Authorization: `Bearer ${token}` },
-        }
-      );
-      alert("게시물 수정 완료.");
-      setEditMode(false);
-      setEditBoardId(null);
-      setTitle("");
-      setContent("");
-      setIsWriting(false);
-      fetchBoards(page, pageSize, category);
+      if (typeof window !== "undefined") {
+        const token = localStorage.getItem("accessToken");
+        await api.put(
+          `/boards/${editBoardId}`,
+          { title, content },
+          {
+            headers: { Authorization: `Bearer ${token}` },
+          }
+        );
+        alert("게시물 수정 완료.");
+        setEditMode(false);
+        setEditBoardId(null);
+        setTitle("");
+        setContent("");
+        setIsWriting(false);
+        fetchBoards(page, pageSize, category);
+      }
     } catch (error: unknown) {
       if (error instanceof AxiosError) {
         console.error("게시판 수정 오류:", error);
@@ -142,17 +150,19 @@ export const useBoards = () => {
     }
   };
 
-  const storedUsername = localStorage.getItem("username");
-  const userId = localStorage.getItem("userId");
-  useEffect(() => {
-    if (storedUsername && userId) {
-      setUser({ username: storedUsername, id: userId });
-    }
-    fetchBoards(1, pageSize, category);
-  }, []);
-  useEffect(() => {
-    fetchBoards(page, pageSize, category);
-  }, [page]);
+  if (typeof window !== "undefined") {
+    const storedUsername = localStorage.getItem("username");
+    const userId = localStorage.getItem("userId");
+    useEffect(() => {
+      if (storedUsername && userId) {
+        setUser({ username: storedUsername, id: userId });
+      }
+      fetchBoards(1, pageSize, category);
+    }, []);
+    useEffect(() => {
+      fetchBoards(page, pageSize, category);
+    }, [page]);
+  }
   return {
     boards,
     fetchBoards,
